+++ RELEASE_NOTES.md
@@ -6,7 +6,7 @@ https://github.com/padogrid
 
 ## Version 0.9.22-SNAPSHOT
 
-### Release Date: 10/26/22
+### Release Date: 11/06/22
 
 - Added support for Confluent Platform. You can now install Confluent Platform by running `install_padogrid -product confluent`. Confluent and Kafka share the same cluster commands. The `CLUSTER_TYPE` value of `kraft` has been replaced with `kafka` and `confluent`. PadoGrid now supports Kafka, Confluent Community, and Confluent Commertial versions.
 - Updated RWE and workspace commands to properly handle file permissions enforced when multitenancy is enabled.
@@ -14,6 +14,7 @@ https://github.com/padogrid
 - Excluded `log4j` from the cluster class paths to prevent version conflicts. Due to the recent security updates made by `log4j`, many of the products are fragile to `log4j` versions other than the versions included in their distributions. PadoGrid now includes `log4j` binary for running client apps only.
 - Fixed Kafka cluster log file name to `kafkaServer.out`. Both `server.log` and `kafkaServer.out` names were used previously.
 - Added bundle support for Kafka and Confluent. See [Kafka Confluent Bundle Catalog](https://github.com/padogrid/catalog-bundles/blob/master/confluent-catalog.md)
+- Added support for viewing Jupyter log files in `show_jupyter`. Use the `-port` option to view the specific server's log file or `-all` to view all active server log files.
 
 ---